+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 1.1 (Status: DEVELOPMENT)
 * New version files.
 * New config parsing.
+* Minor bug/spell fixes in master aswell as slave config changed (section name)
 
 Version 1.0.1-stable (Status: RELEASED)
 * Slave: Fixed serious writing issue where ' is used instead of ".